Windows native fixes

+++ include/cute/arch/xe_copy_1B.hpp
@@ -231,7 +231,7 @@ struct XE_2D_U8x1x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<ushort *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m1k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -249,7 +249,7 @@ struct XE_2D_U8x2x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m2k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -266,7 +266,7 @@ struct XE_2D_U8x2x32_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u16_m2k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::ushort2 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -285,7 +285,7 @@ struct XE_2D_U8x4x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m4k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -303,7 +303,7 @@ struct XE_2D_U8x8x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m8k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -321,7 +321,7 @@ struct XE_2D_U8x16x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m16k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -333,7 +333,7 @@ struct XE_2D_U8x16x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -354,7 +354,7 @@ struct XE_2D_U8x32x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m32k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -372,7 +372,7 @@ struct XE_2D_U8x1x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m1k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -404,7 +404,7 @@ struct XE_2D_U8x2x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m2k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -436,7 +436,7 @@ struct XE_2D_U8x4x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m4k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -468,7 +468,7 @@ struct XE_2D_U8x8x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m8k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -500,7 +500,7 @@ struct XE_2D_U8x16x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m16k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -512,7 +512,7 @@ struct XE_2D_U8x16x64_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -533,7 +533,7 @@ struct XE_2D_U8x32x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort64 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m32k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -545,7 +545,7 @@ struct XE_2D_U8x32x64_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m32k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -568,7 +568,7 @@ struct XE_2D_U8x32x16_LD_V {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::uint8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u8_k32(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -600,7 +600,7 @@ struct XE_2D_U8x32x32_LD_V {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::uint16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u8_k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -618,7 +618,7 @@ struct XE_2D_U8x32x64_LD_V {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::uint32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u8_k32v4(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -635,7 +635,7 @@ struct XE_2D_U8x1x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u8_m1k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::uchar *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -653,7 +653,7 @@ struct XE_2D_U8x2x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u8_m2k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::uchar2 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -671,7 +671,7 @@ struct XE_2D_U8x4x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u8_m4k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::uchar4 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -687,7 +687,7 @@ struct XE_2D_U8x8x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u8_m8k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::uchar8 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -703,7 +703,7 @@ struct XE_2D_U8x8x32_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     __builtin_IB_subgroup_block_write_flat_u8_m8k16v2(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::uchar8 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
