+++ include/cute/arch/xe_copy_2B.hpp
@@ -265,7 +265,7 @@ struct XE_2D_U16x1x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<ushort *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m1k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -283,7 +283,7 @@ struct XE_2D_U16x2x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m2k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -301,7 +301,7 @@ struct XE_2D_U16x4x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m4k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -319,7 +319,7 @@ struct XE_2D_U16x8x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m8k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -331,7 +331,7 @@ struct XE_2D_U16x8x16_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m8k16v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -352,7 +352,7 @@ struct XE_2D_U16x16x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m16k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -364,7 +364,7 @@ struct XE_2D_U16x16x16_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -385,7 +385,7 @@ struct XE_2D_U16x32x16_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m32k16v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -397,7 +397,7 @@ struct XE_2D_U16x32x16_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m32k16v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -418,7 +418,7 @@ struct XE_2D_U16x1x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m1k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -450,7 +450,7 @@ struct XE_2D_U16x2x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m2k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -482,7 +482,7 @@ struct XE_2D_U16x4x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m4k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -514,7 +514,7 @@ struct XE_2D_U16x8x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m8k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -526,7 +526,7 @@ struct XE_2D_U16x8x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m8k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -547,7 +547,7 @@ struct XE_2D_U16x16x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m16k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -559,7 +559,7 @@ struct XE_2D_U16x16x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -580,7 +580,7 @@ struct XE_2D_U16x32x32_LD_N {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::ushort64 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u16_m32k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -593,7 +593,7 @@ struct XE_2D_U16x32x32_LD_N {
 #if defined(SYCL_INTEL_TARGET)
       // __builtin_IB_subgroup_block_read_prefetch_u16_m32k16v2(
       __builtin_IB_subgroup_block_read_prefetch_u16_m8k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -614,7 +614,7 @@ struct XE_2D_U16x16x16_LD_V {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::uint8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u16_k16(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -626,7 +626,7 @@ struct XE_2D_U16x16x16_LD_V {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -647,7 +647,7 @@ struct XE_2D_U16x32x16_LD_V {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::uint16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u16_k32(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -659,7 +659,7 @@ struct XE_2D_U16x32x16_LD_V {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m32k16v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -680,7 +680,7 @@ struct XE_2D_U16x16x32_LD_V {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::uint16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u16_k16v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -692,7 +692,7 @@ struct XE_2D_U16x16x32_LD_V {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -713,7 +713,7 @@ struct XE_2D_U16x32x32_LD_V {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::uint32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transform_u16_k32v2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -725,7 +725,7 @@ struct XE_2D_U16x32x32_LD_V {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u16_m16k16v2(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -749,7 +749,7 @@ struct XE_2D_U16x16x8_LD_T {
     static_assert(sizeof(T) == 2, "Expected T to have size 4");
     *reinterpret_cast<intel::uint4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transpose_u32_k4(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -770,7 +770,7 @@ struct XE_2D_U16x16x16_LD_T {
     static_assert(sizeof(T) == 2, "Expected T to have size 2");
     *reinterpret_cast<intel::uint8 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transpose_u32_k8(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -787,7 +787,7 @@ struct XE_2D_U16x1x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     // static_assert(sizeof(T) == 2, "Expected T to have size 2");
     __builtin_IB_subgroup_block_write_flat_u16_m1k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(ushort *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -805,7 +805,7 @@ struct XE_2D_U16x2x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     // static_assert(sizeof(T) == 2, "Expected T to have size 2");
     __builtin_IB_subgroup_block_write_flat_u16_m2k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::ushort2 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -823,7 +823,7 @@ struct XE_2D_U16x4x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     // static_assert(sizeof(T) == 2, "Expected T to have size 2");
     __builtin_IB_subgroup_block_write_flat_u16_m4k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::ushort4 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
@@ -841,7 +841,7 @@ struct XE_2D_U16x8x16_ST_N {
 #if defined(SYCL_INTEL_TARGET)
     // static_assert(sizeof(T) == 2, "Expected T to have size 2");
     __builtin_IB_subgroup_block_write_flat_u16_m8k16v1(
-        (long)(baseoffset), width - 1, height - 1, pitch - 1, coord,
+        (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord,
         *(intel::ushort8 *)(src));
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
